ogg: rationalize serialno type to guint32
It is a 32 bit unsigned number.
Sure, the libogg API uses a long, but that's an unfortunate oversight.

https://bugzilla.gnome.org/show_bug.cgi?id=656775

+++ ext/ogg/gstoggstream.h
@@ -45,7 +45,7 @@ struct _GstOggStream
 {
   ogg_stream_state stream;
 
-  glong serialno;
+  guint32 serialno;
   GList *headers;
   gboolean have_headers;
   GList *queued;
